pyproject.toml
[project]
name = "dbx-finance"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "databricks-connect>=16.1.5",
    "databricks-sdk>=0.56.0",
    "delta-spark>=3.3.2",
    "pandas>=2.3.0",
    "pyspark>=3.5.6",
    "yfinance>=0.2.61",
]
